rename group.api to namespace.api to match naming in renku-data-services
